Inform submitter of merge failure by dialog box
This simplifies identifying a submit failure, because we now show
the message in a popup dialog box, so the user doesn't need to scroll
down to the bottom of the page to view the most recent message.

We also cut down one RPC during submit and abandon by returning
the updated change as the result of the submit or the abandon RPC.
This way the client has the new change state in memory and can
immediately redraw the screen.

Change-Id: I471746e1398ef3be1e4cb07e333ab01141256ad6
Signed-off-by: Shawn O. Pearce <sop@google.com>

Former-commit-id: 3d2f85006a5c045a5b3748b1cbb1268f25846227