Move git pull URL to bottom of email notifications
Andy pointed out that Gmail was making the git pull URL a hyperlink,
which made clicking the review URL error prone as the two targets
were really close to each other.  By moving the git pull URL to the
bottom of the email, just above the footer, we make it less likely
that the target is confused.

Change-Id: Ibd03e4f9b3fa9d54358a2f2ec2853c61e3e00b48
Signed-off-by: Shawn O. Pearce <sop@google.com>

Former-commit-id: b3645d5654ed9a42b19644a4d8f503e2a9469ad0