Switch to the new org.eclipse.jgit package
Adds support for core.logAllRefUpdates, and also moves us to the new
org.eclipse.jgit namespace, which is the future of the JGit project.

Unfortunately this version is a pre-release version which comes
with some risk that the project history will be rewritten, possibly
allowing this version to become lost in time.

Change-Id: I780d2b3f0af66f32f31f74ff8f3fa14fdaad721d
Signed-off-by: Shawn O. Pearce <sop@google.com>

Former-commit-id: 1e1589c9a1a82c27f686cc3d622aee3a32be9946