Allow admins to search for and bind to LDAP groups
Group owners may now search the LDAP directory for matching group
names, and select one to bind their newly created Gerrit group to.

This is a very crude interface, but makes it a lot easier to define
a new LDAP group from within the web UI.  A group owner can create
a new group, change its type to "LDAP Group", and then query the
local LDAP directory for groups whose name match a search filter.
Once a matching record is found, selecting the record will save
the DN into the group object in the Gerrit database.

Unfortunately the LDAP group search is performed using the Gerrit
server's LDAP credentials, and not those of the user who is
performing the search.  This may cause group names to be leaked to
parties who should not know them.  To reduce the exposure of the
directory, only the DN is displayed in the search results.

Bug: issue 297
Change-Id: Ia90f21dc8fbc55216e5c9d1f8159a5f84130f2a7
Signed-off-by: Shawn O. Pearce <sop@google.com>

Former-commit-id: f0125d734edaa270b0498bb603010c6f4f89f60a