Support SMTP over SSL/TLS
If sendemail.smtpEncryption is set to 'ssl' or 'tls' we now enable
the proper encryption on top of the socket, permitting the client
to use encryption when talking with the relay SMTP server.
This might be necessary to protect the username/password used to
authenticate prior to sending a message.

Bug: issue 300
Change-Id: Idecb20326261cbc8951e2ff469b95ea7ee83e48c
Signed-off-by: Shawn O. Pearce <sop@google.com>

Former-commit-id: 6e9a83f6ca5513dbb3729b153b14a8129cf74feb