Grant anonymous uses READ +1, registered users READ +1..+2
The READ +2 permission implies being able to read a project and
uploading a new change into it for review.  Only a registered user
can actually do an upload, as Gerrit requires the user to have an
account that is associated with the change review.

Therefore in our initial configuration of "-- All Projects --"
access rights we should grant read access to anonymous users but
read + create change for review access for registered users.

Change-Id: I57ecd17cfabc093b1dfd068a4f040dd067e5f5c0
Signed-off-by: Shawn O. Pearce <sop@google.com>

Former-commit-id: 04f7c92b3e12740d35c836c02d6084fd874c5632