Fix auto-formatting of bullet lists
Previously we were bulleting the first line in a paragraph, even if
it did not have a bullet marker.  E.g. if the user entered:

  To see this bug, you have to:
  * Be on IMAP or EAS (not on POP)

the first line was treated as though it had been "* To see this bug".
This was not what the author intended in this example.

We now make an exception for the first line(s) in the list paragraph,
rendering them in their own <p> tag isolated from the list elements.

Bug: gerrit issue 293
Change-Id: I7131bb845d2eeb9e208300ba9a09ec6c805b8df8
Signed-off-by: Shawn O. Pearce <sop@google.com>

Former-commit-id: 8f608ec926c3418d7c84f35bd0ef371acf5d18d0