Forbid use of anonymous servlets on any container
The /servlet/* mapping is used by the servlet specification to mean
invoking any servlet by their dynamic class name.  This might be a
security risk to the application as it could allow a user to invoke
some servlet we have specifically not mapped into our URI space as
we have chosen not to enable its functionality in this instance.

Fortunately all popular containers, including Jetty and Tomcat have
this mapping disabled by default.  However it can be enabled as a
container-wide feature in both products and if we are co-hosted with
another application in a container which has enabled this feature we
might start seeing dynamic servlet requests.

Instead remap them all to a 404 Not Found error response.

Change-Id: Ib96d4b76b9d226d9248b2334131d5b8912b34c52
Signed-off-by: Shawn O. Pearce <sop@google.com>

Former-commit-id: 58120ab4e9df0e95289115f600928ac1790029e8