Use listeners to manage server startup/shutdown
Instead of enumerating the startup/shutdown sequence inside of the
WebAppInitializer we now use listeners which are bound and managed
by Guice.  The listeners are fired in the order they are registered
within the injector modules.

By using Guice we are more easily able to track the need to start
(or gracefully stop) a component alongside its explicit binding
in the injector.  We can also conditionally include start or stop
rules by controlling which modules are included in the injection.

Change-Id: I93590c666d46e13fdce9aa05100489f9f6d94615
Signed-off-by: Shawn O. Pearce <sop@google.com>

Former-commit-id: 17a30c3703fda4026a1b6633f5bddb3ed3c36b74