Support invoking gitweb from within Gerrit
By running gitweb within Gerrit we can apply project access control
rules before gitweb gets a chance to execute.  This allows us to
hide a project from gitweb that the current user isn't permitted
to access.

We also can simplify some of the configuration for gitweb by making
our own configuration file with settings that are appropriate for this
particular Gerrit installation and the git repositories it manages.

Bug: issue 183
Change-Id: Ic5739d0b8e52fb842f3d132300b58aa8b25bfdf6
Signed-off-by: Shawn O. Pearce <sop@google.com>

Former-commit-id: b906a3e22fdb636a9e5b0eeb3ec82102a7c82d92