Use JGit's http_proxy based initialization for the J2SE
libcurl honors the http_proxy environment variable, so it might be
reasonable for this to be set in the user's environment if they
are running on a POSIX system.  We should try to use anything
contained here for any network connections we create out of the
J2SE URL library.

Bug: issue 323
Change-Id: Id725e991f1e820f0d138ab659d01949dbe261852
Signed-off-by: Shawn O. Pearce <sop@google.com>

Former-commit-id: da85ff6cb4ecd37dfa6ebe60658502577ecd4d84