gsql: Fix \d on H2
H2 has a quirk in its getTables ResultSet where the schema name
is stored in a column called "TABLE_SCHEMA" not "TABLE_SCHEM".
This threw off our poor-man's version of ResultSet.findColumn(),
which the H2 driver implements to handle both names.

H2 also stores table names all uppercase, not all lowercase.
We need to adjust the table name as supplied by the user to match
the style the driver says the database stores.

If a table doesn't exist, print an error message instead of printing
out an empty column list.

Bug: issue 341
Change-Id: Ia7d0cb74ccabc44ed58d13f602ee7f95883d5586
Signed-off-by: Shawn O. Pearce <sop@google.com>

Former-commit-id: e926fa7c0b047ff69fe56148497ab38937d1b533