gsql: Improve formatting of column types and indexes
Apparently the column type data isn't complete unless we pull several
different fields from the getColumns ResultSet and mix it together.
So instead of showing a handful of columns from the ResultSet we mix
them together with a computation function.

The index data is also scattered, each index spans multiple rows in
the ResultSet, one for each column of the index.  Since many of the
indexes in Gerrit's schema are compound we collect these into one
object per index and then format the results after the ResultSet has
been fully traversed.

Change-Id: I490aa2eede3323cda15e025f0f16f8a3d69c7e0e
Signed-off-by: Shawn O. Pearce <sop@google.com>

Former-commit-id: 036431240554cb8d8dc565737537901b23996c17