Always trim Change-Id lines to handle non-standard whitespace
A user might copy-and-paste a Change-Id line from the little info
table on the web, which usually brings in a tab rather than a space
between the tag and the identity.  So we would see [1] instead of
seeing the more typical string [2].

 [1] "Change-Id:\tIdeadbeef...\n"
 [2] "Change-Id: Ideadbeef...\n".

Prior to this bug fix we were trimming during insertion of a new
change into the database, but we did not trim when querying to
look for replacement or merge candidates.  This could lead the same
Change-Id showing up twice in the database.

Change-Id: Ia90b894eda545abf0b976b1610ae68bb4147d963
Signed-off-by: Shawn O. Pearce <sop@google.com>

Former-commit-id: 1a9029b006484e8b9aca06ff261beb2324bb9916