Don't enable replication if replication.config is empty
Just like if the file doesn't exist, an empty file implies there
is no replication configured and we should skip processing.  Log
an informational message so the administrator knows the file is
present and being read, but hasn't been configured yet.

Explicitly checking for this and logging is a good idea because
init is now creating the replication.config file anytime it sets
up a site path.

Change-Id: Iae9e8f47fb688ec05b3b8b5dc28e44d4df1ad419
Signed-off-by: Shawn O. Pearce <sop@google.com>

Former-commit-id: feabf4cd74f67938896e5d4e8c8df88740734ed7