init: Initialize system_config.site_path
We failed to set site_path in the system_config table, because it
usually isn't necessary for init and daemon.

Currently, site_path is only required when loading within a servlet
container where the database environment is handed down via JNDI
but the site_path isn't available.

For our own command line tools, site_path is read from the command
line options and the database's stored value is ignored and not
required for execution.

Unfortunately we have a bit of a chicken-and-egg problem here as
site_path isn't available when we are creating the initial database
record in the system_config table.  So we have to go back and update
it after the fact.

Change-Id: I3683014f953178e019ebd65625a3c85cc5802403
Signed-off-by: Shawn O. Pearce <sop@google.com>

Former-commit-id: 4b798a105ff6a4c331b4af3de5356d0812f0b099