Add a Register link when using LDAP authentication
If auth.type is LDAP Gerrit is responsible for collecting the username
and password and authenticating to the directory.  But we don't have a
way to create new users in the directory, as this is far beyond any
scope we want to handle.

Site administrators might have a URL users can visit to create their
own account, or maybe just want to publish a mailto: link so the user
can request an account to be manually created for them.  Allow those
links to be configured through gerrit.config as auth.registerUrl.

Change-Id: I996ed53c5b80579f3a60cfcf9aebb7940d501366
Signed-off-by: Shawn O. Pearce <sop@google.com>

Former-commit-id: c9d26b53363fe78827bce92269113ebacf0497f0