Remove legacy tools/to_jetty.sh
This only existed for developers to bundle their WAR and move
it into their external Jetty container for testing.  Now that
we have our own internal daemon, we really prefer this for any
development work because its self-contained and starts up faster.
Its also implemented by Jetty and is easier to upgrade, making it
more likely we'll be running current stable, patched code.

Change-Id: I00ce142c9ca1009ff49e3ff70f3c977d3c50866e
Signed-off-by: Shawn O. Pearce <sop@google.com>

Former-commit-id: eda42398d1696c2dec366e3b5e42255a908cc131