CopyableLabel: Allow setting the preview text for the label
When Gerrit Code Review puts the entire SSH host key into the
copyable label it causes the entire layout to fall apart as
the text is simply too wide for the screen, and the Clippy
Flash movie runs off the edge.

Instead allow the application to provide its own abbreviation
string for the preview you click on.

Change-Id: I5aaaa7b4b41c2f461671955a7d82f09107e0e0a5
Signed-off-by: Shawn O. Pearce <sop@google.com>

Former-commit-id: 880abba733308d68882c07cdf9296ab2fedf62ec