Clean up the DWIMery for database.* configuration settings
This code was a mess.  We went through a lot of hoops to try and
figure out what the administrator wanted from us given strange
combinations of database.type and database.url.  Instead use a
very simple set of rules that pretty much requires database.type
or database.url to be set to indicate how we should proceed.

Change-Id: I44ef20592df288616907d967b46c87fecc9ed80c
Signed-off-by: Shawn O. Pearce <sop@google.com>

Former-commit-id: 1be39064f6ab151487c7a6820b8789785f587178