Support creating new users in DEVELOPMENT_BECOME_ANY_ACCOUNT
Under development mode it can be useful to construct a new user
account anytime, without needing to manually inject records in
the database or switch temporarily to LDAP or OpenID methods.

Because we don't really have an external identity for the user
we given them a random one via the JRE's UUID generator.  New
account creation should be infrequent enough in development to
never run into a collision.

Change-Id: Id15ddc34ef7bfa5cea6d12c435cb09b8443c77fd
Signed-off-by: Shawn O. Pearce <sop@google.com>

Former-commit-id: f2d68b182d95877477481817bf466527e214ad9a