Make gitweb url links customizable, add support for cgit
The gitweb.type variable can now be one of the following: "gitweb",
"cgit" or "custom", with suitable defaults for other variables
selected based on the preferred URL format for that gitweb type.

If gitweb.type = custom is used, then URL patterns must be defined
in the gerrit.config file:

  [gitweb]
    project = ${project}/summary
    revision = ${project}/commit/${commit}
    branch = ${project}/log/?h=${branch}

In a custom configuration, all 3 variables must be defined or gitweb
integration will be disabled.

Bug: issue 366
Bug: issue 371
Change-Id: Ic681af1300b9c7cbae1f0f0ce31a5272378feac6
Signed-off-by: Shawn O. Pearce <sop@google.com>

Former-commit-id: 27868a42d9b6aca08c0588eafa55579a2bf81650