Drop the lastUsedOn fields from AccountSshKeys, AccountExternalIds
By dropping this column we avoid updating the database during the
authentication phase for a new web session, or a new SSH session.

Avoiding this update on read-only slaves can improve performance if
they have very slow connectivity to the master database.  On the
master node, avoiding the update saves a few relatively useless
disk IOs on the database.

This change also removes one of the issues we had with converting
the user account schema to GiMD.  Updating the account database
in Git every time a user authenticates would create thousands of
objects per hour on busy servers, all of them almost completely
useless to the system administrator.

This change makes a trade-off of faster session authentication, in
return for a bit less information about account usage.  Instead of
pulling account usage from the database, we'll have to start writing
to local log files instead.

Change-Id: Ie335f5e6e0f5ff16b987d4e2698bc25ff14bd989
Signed-off-by: Shawn O. Pearce <sop@google.com>

Former-commit-id: 5f95a93f0f86258c8cbc60a528eec96ad90ce292